+++ docs/CHANGELOG.md
@@ -6,12 +6,16 @@
 - frontend: improved text-editor + pdf-support
 - frontend: fixed album-filename-restrictions
 - frontend: added support for iframe-print-preview for pdf in android-aps ;-)
+- frontend: added changelog-component
+- frontend: show changelog in page-management
 
 ### new features
 - frontend: added support for iframe-print-preview for pdf in android-aps ;-)
+- frontend: added changelog-component
 
 ### improvements
 - frontend: improved text-editor + pdf-support
+- frontend: show changelog in page-management
 
 ### bug fixes
 - frontend: fixed album-filename-restrictions